+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [BREAKING] Fix parameter order of set person home/away methods
+- Refactor camera person detection checks
 
 ### Deprecated